pyproject.toml
[tool.poetry]
name = "webvep"
version = "0.1.0"
description = "API and frontend for VEP tool"
authors = ["Ian Vermes <ian.vermes@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7"
django = ">=2.2"
psycopg2-binary = "*"
pytest = "*"
djangorestframework = "^3.11.0"
markdown = "^3.2.1"
django-filter = "^2.2.0"
pytest-docker-compose = "^3.1.2"
vcfpy = "^0.12.1"
pysam = "^0.15.4"

[tool.poetry.dev-dependencies]
ipython = "*"
django-stubs = "^1.1.0"
mypy = "^0.740"
black = "^19.10b0"
pytest-django = "^3.8.0"
pytest-pythonpath = "^0.7.3"
bs4 = "^0.0.1"
pytest-docker-compose = "^3.1.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
